Remove runners other than direct
